refactor(review): extract FindingPipeline from ReviewOrchestrator (#250)
Step 7 of the decomposition: move the post-AI finding chain into a
FindingPipeline.refine() — quote validation, dedupe, verification against
the diff, drop-already-replied, missing-anchor backfill, and persistence.
Ordering preserved verbatim (validate before dedupe). The DiffLineResolver
is shared with the verdict backstop, so review() still builds it and
threads it into refine().

quoteValidator, deduplicator, findingVerificationService and the
ObjectMapper leave the orchestrator (followUpAnalyzer stays shared),
dropping it to 16 injected collaborators and ~415 LOC — review() is now a
readable ordered sequence of phase calls.

Tests: the orchestrator test builds a real FindingPipeline from the
mocked collaborators; the persistAiResponse/populateMissingAnchors cases
call it directly (the serialization-failure case builds a pipeline with a
throwing ObjectMapper). Full suite green, spotbugs + spotless clean.
